DAT-368 SDN4 prep work - Backoff policies, moving Node to Cypher

+++ fd-common/src/main/java/org/flockdata/track/model/EntityTag.java
@@ -55,7 +55,7 @@ public interface EntityTag {
      *
      * @return weight of the attribute in the relationship
      */
-    Integer getWeight();
+    Long getWeight();
 
     public GeoData getGeoData();
